requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
plotly>=5.0.0
langchain==0.1.14
langchain-community==0.0.30
langchain-chroma==0.1.0
chroma-db==0.4.22
langchainhub==0.1.14
langgraph==0.0.44
crewai==0.25.0
llama-index==0.9.48
llama-index-embeddings-huggingface==0.2.0
llama-index-llms-local==0.1.0
ollama==0.0.60
rank-bm25==0.2.2
numpy==1.24.3
pandas==2.0.3
pydantic==2.3.0
python-dotenv==1.0.0
pytest==7.4.3
ragas==0.1.6
datasets==2.14.6
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
redis==5.0.0
supabase==2.4.1
python-dateutil==2.8.2
python>=3.12,<3.13
